Use different classes for different aggregations
